Setup the route for clicked collection & Refactor components arrangement
