+++ src/components/Banner.js
@@ -1,10 +1,11 @@
 import React from 'react';
 
-const HomeBanner = () => {
+const HomeBanner = (props) => {
   console.log('home banner');
+  const { title } = props;
   return (
     <div className="home-banner col-12 d-flex flex-column justify-content-center">
-      <h1 className="fa align-self-center border border-warning border-3 rounded p-3 banner-title">Photo Collections Gallery</h1>
+      <h1 className="fa align-self-center border border-warning border-3 rounded p-3 banner-title">{title}</h1>
     </div>
   );
 };
