+++ src/components/target_collection/TargetCollection.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Banner from '../Banner';
+
+const TargetCollection = () => (
+  <div className="">
+    <Banner title="collection" />
+    Target Collection
+  </div>
+);
+
+export default TargetCollection;
